apps/ml/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
weaviate-client>=4.4.0
psycopg2-binary>=2.9.9
sentence-transformers>=2.3.1
openai>=1.10.0
ollama>=0.1.6
groq>=0.4.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
langchain-text-splitters>=0.0.1
